Added key, door and scene transition
Added key, door and scene transition scripts. Keys can be picked up, doors and can be interacted with by running into them and tp the player to a connected door. The scene transition works if the player runs into the ending door of the level. Also added some basic box colliders to the platforms and objects of level 1, this could be improved. Also fixed a couple bugs with the player, namely the collider not matching the sprite and it still had the eye from the prototype object.